requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.0
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.27.0
config==0.5.1
decouple==0.0.7
demoApp==0.0.1
dj-database-url==0.5.0
Django==3.2.9
django-celery-results==2.2.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
Faker==9.8.0
gunicorn==20.1.0
kombu==5.2.1
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==3.2.1
prompt-toolkit==3.0.22
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
UploadView==0.2.7
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
WebOb==1.8.7
whitenoise==5.3.0
